Task: add time formatting to messages
