+++ app/src/config/index.js
@@ -1,1 +1,2 @@
 export const HOST = 'http://localhost:5001/api/';
+export const DateFormat = 'MM/DD/YY h:mm';
